backend/requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
openai
python-dotenv
python-jose
passlib
python-multipart
httpx
google-auth
psycopg2-binary
tiktoken
email-validator
bcrypt